server/backend/requirements.txt
django==4.1.7
sqlparse>=0.2.2
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
mysqlclient==2.1.1
regex==2022.10.31
Pyrebase4==4.5.0
django-redis==5.2.0
channels==3.0.4 
channels-redis==4.0.0
channels-rabbitmq==4.0.1
requests==2.28.2
requests-toolbelt==0.10.1